+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Changelog History
 =================
 
+Intake-esm v2019.5.?? (2019-05-??)
+===================================
+
+
+Features
+---------
+
+- Add implementation for The Gridded Meteorological Ensemble Tool (GMET) data holdings (:pr:`61`) `Anderson Banihirwe`_
+
+
+
 Intake-esm v2019.4.26 (2019-04-26)
 ===================================